requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.1.0
bcrypt==4.3.0
boto3==1.40.4
botocore==1.40.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
fastapi-cli==0.0.7
gotrue==2.12.3
greenlet==3.2.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.6
openai==1.98.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.3.0
pluggy==1.5.0
postgrest==1.1.1
psycopg==3.2.9
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
realtime==2.6.0
requests==2.31.0
rich==14.0.0
rich-toolkit==0.14.6
s3transfer==0.13.1
selenium==4.29.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
starlette==0.46.2
storage3==0.12.0
StrEnum==0.4.15
supabase==2.17.0
supafunc==0.10.1
tomli==2.2.1
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.1
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
